% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab.gplot.factor.R
\name{ab.gplot.factor}
\alias{ab.gplot.factor}
\title{Title}
\usage{
ab.gplot.factor(x, y, se.y = NULL, xlab, ylab, residuals = NULL,
  rugplot = FALSE, scale = 0, se = FALSE, xlim = NULL, ylim = NULL,
  fit = TRUE, ...)
}
\arguments{
\item{x}{}

\item{y}{}

\item{se.y}{}

\item{xlab}{}

\item{ylab}{}

\item{residuals}{}

\item{rugplot}{}

\item{scale}{}

\item{se}{}

\item{xlim}{}

\item{ylim}{}

\item{fit}{}

\item{...}{}
}
\description{
Title
}
