% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.gam.replace.R
\name{ab.na.gam.replace}
\alias{ab.na.gam.replace}
\title{Missing Data Filter for GAMs}
\usage{
ab.na.gam.replace(frame)
}
\arguments{
\item{frame}{data frame}
}
\description{
Cambiato leggermente la funzione na.gam.replace perchè possa essere utilizzata con i data.frame
}
\details{
\code{\link{na.gam.replace}}
}
