% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab.plot.gam.R
\name{ab.plot.gam}
\alias{ab.plot.gam}
\title{ab.plot.gam}
\usage{
ab.plot.gam(x, residuals = NULL, rugplot = TRUE, se = FALSE, scale = 0,
  ask = FALSE, terms = labels.gam(x), ...)
}
\arguments{
\item{x}{}

\item{residuals}{}

\item{rugplot}{}

\item{se}{}

\item{scale}{}

\item{ask}{}

\item{terms}{}

\item{...}{}
}
\description{
ab.plot.gam
}
