% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab.plot.preplot.gam.R
\name{ab.plot.preplot.gam}
\alias{ab.plot.preplot.gam}
\title{Title}
\usage{
ab.plot.preplot.gam(x, y = NULL, residuals = NULL, rugplot = TRUE,
  se = FALSE, scale = 0, fit = TRUE, ...)
}
\arguments{
\item{x}{}

\item{y}{}

\item{residuals}{}

\item{rugplot}{}

\item{se}{}

\item{scale}{}

\item{fit}{}

\item{...}{}
}
\description{
Title
}
