% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ada_crossvalidation.R
\name{ada_crossvalidation}
\alias{ada_crossvalidation}
\title{adaboost cv}
\usage{
ada_crossvalidation(i, formula = NULL, y, x, folds, iter = 200)
}
\arguments{
\item{i}{i-esimo fold}

\item{formula}{formula del modello}

\item{y}{vettore. variabile risposta}

\item{x}{matrice variabili esplicative}

\item{folds}{prodotto da kfold function}

\item{iter}{numero iterazioni}
}
\description{
fonisce le previsioni della cv, da utilizzare dentro sapply(1:k,...)
}
