% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance.R
\name{balance}
\alias{balance}
\alias{balance.data.table}
\alias{balance.data.frame}
\title{Bilanciare}
\usage{
balance(data, y = "y")

\method{balance}{data.table}(data, y = "y")

\method{balance}{data.frame}(data, y = "y")
}
\description{
Bilanciare
}
