% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bastoncini.R
\name{bastoncini}
\alias{bastoncini}
\title{diagramma a bastoncini}
\usage{
bastoncini(x, perc = TRUE, nome = NULL)
}
\arguments{
\item{x}{vettore}

\item{perc}{}

\item{nome}{}
}
\description{
diagramma a bastoncini
}
