% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_cat.R
\name{change_cat}
\alias{change_cat}
\title{Sostituire categorie}
\usage{
change_cat(data, old = NULL, new = "other", condition = 1)
}
\arguments{
\item{data}{data.frame o vettore}

\item{old}{le categorie da sostituire}

\item{new}{la new da inserire}
}
\description{
sostituisce categorie con un'altra
}
