% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_list.R
\name{create_list}
\alias{create_list}
\title{crea liste}
\usage{
create_list(iniziale, id, ...)
}
\arguments{
\item{iniziale}{}

\item{id}{}

\item{...}{altri elementi da aggiungere alla lista}
}
\description{
crea una lista con oggetti che iniziano per una stessa iniziale, ad es. 'm', e poi hanno un numero identificativo (es.'1:16')
}
