% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.knn.R
\name{cv.knn}
\alias{cv.knn}
\title{cross validation leave one out knn}
\usage{
cv.knn(K, x, y, folds = NULL)
}
\arguments{
\item{K}{}

\item{x}{}

\item{y}{}

\item{folds}{}
}
\value{
errore totale per un determinato k
}
\description{
cross validation leave one out knn
}
