% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.pc.R
\name{cv.pc}
\alias{cv.pc}
\title{cv regressione componenti principali}
\usage{
cv.pc(j, y, x, folds, family = binomial, cor = T)
}
\arguments{
\item{y}{vettore. variabile risposta}

\item{x}{matrice variabili esplicative}

\item{folds}{prodotto da kfold function}

\item{family}{qualsiasi}

\item{cor}{logical. if correlation or covariance}

\item{i}{i-esimo fold}
}
\description{
fonisce le previsioni della cv, da utilizzare dentro sapply(1:k,...)
}
