% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elim_cat.R
\name{elim_cat}
\alias{elim_cat}
\title{elimina categoria}
\usage{
elim_cat(data, cat)
}
\arguments{
\item{data}{data.frame}

\item{cat}{vector, allora cerca elimina la categoria da tutto il dataset, lista e allora elimina categorie specifiche per ogni colonna}
}
\value{
il dataset modificato
}
\description{
mette i valori corrispondenti a quella categoria a NA, elimina poi la categoria vuota
}
