% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estrai_cat.R
\name{estrai_cat}
\alias{estrai_cat}
\title{estrarre categorie}
\usage{
estrai_cat(data, nstring, label = NULL)
}
\arguments{
\item{data}{vettore}

\item{nstring}{vettore se si deve estrarre una sola variabile, lista per estrarre più variabili.}
}
\description{
estrarre categorie
}
