% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factordummy.R
\name{factordummy}
\alias{factordummy}
\title{fattorizza le dummy}
\usage{
factordummy(data)
}
\arguments{
\item{data}{data.frame}
}
\description{
cerca le variabili che hanno valori unici c(0,1) e mette quelle variabili a fattori
}
