% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_gam.R
\name{formula_gam}
\alias{formula_gam}
\title{Formula per gam}
\usage{
formula_gam(x, y = NULL, kdf = NULL, formula_y = "as.factor(y)")
}
\arguments{
\item{k}{}
}
\description{
crea una formula additiva dove lascia inalterato i fattori, le altre le racchiude tra s().
}
