% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward cv.R
\name{forward_crossvalidation}
\alias{forward_crossvalidation}
\title{Forward cross validation}
\usage{
forward_crossvalidation(i, formula = NULL, data, folds, family = binomial)
}
\arguments{
\item{i}{}

\item{formula}{}

\item{data}{}

\item{folds}{}

\item{family}{}
}
\description{
Forward cross validation
}
