% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.R
\name{freq}
\alias{freq}
\title{Grafici a barre}
\usage{
freq(data, subset)
}
\arguments{
\item{data}{data.frame: la variabile risposta deve y}

\item{subset}{colonne delle esplicative selezionate}
}
\value{
a list: le tabelle di frequenza con cui ha fatto i grafici
}
\description{
Puo' disegnare 2 grafici a barre, ho riportare le rispettive tabelle di frequenza.
}
