% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequenze.R
\name{frequenze}
\alias{frequenze}
\title{frequenze}
\usage{
frequenze(x, cumul = FALSE, head = NULL, reverse = F)
}
\arguments{
\item{x}{vettore}

\item{cumul}{se vuoi le frequenze cumulate}

\item{head}{vettore delle frequenze che si vuol selezionare}

\item{reverse}{dal più piccolo al più grande}
}
\description{
stampa una tabella con le frequenze di ogni singola modalita'
}
