% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_crossvalidation.R
\name{gam_crossvalidation}
\alias{gam_crossvalidation}
\title{Gam cv}
\usage{
gam_crossvalidation(i, x, folds, family = binomial)
}
\arguments{
\item{x}{}

\item{folds}{}

\item{family}{}

\item{formula}{}
}
\value{
predictions
}
\description{
funzione da inserire in sapply per la cv
}
