% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputazione.R
\name{imputazione}
\alias{imputazione}
\title{Imputazione}
\usage{
imputazione(object, data, exclude = NULL)
}
\arguments{
\item{object}{list. I modelli concatenati in una lista. Devono avere lo stesso ordine di \code{unique(search_NA(data,exclude=exclude))}}

\item{data}{data.frame}

\item{exclude}{stesso di search_NA}
}
\value{
dataset riempito
}
\description{
Sostituisce gli NA con i valori previsti dalla regressione
}
