% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.liftroc.R
\name{join.liftroc}
\alias{join.liftroc}
\title{join.liftroc}
\usage{
join.liftroc(g, legend = NULL)
}
\arguments{
\item{g}{lista dei grafici da unire}

\item{legend}{vettore con i nomi della legenda}
}
\description{
funzione che unisce in un unico grafico tutte le curve lift e tutte le curve ROC
}
