% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_fun.R
\name{label_fun}
\alias{label_fun}
\title{label_fun}
\usage{
label_fun(fit, variabili = NULL, ...)
}
\arguments{
\item{fit}{glmnet}

\item{variabili}{vettore di caratteri delle variabili da scrivere}

\item{...}{argomenti di \code{\link{text}}}
}
\description{
label_fun
}
