% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lift_roc.R
\name{lift_roc}
\alias{lift_roc}
\title{Lift e roc curve}
\usage{
lift_roc(previsti, g, type = "bin", plot.it = TRUE, soglia = 0.5)
}
\arguments{
\item{previsti}{mettere le probabilità di 1}

\item{g}{la variabile risposta}

\item{type}{crude, bin o sm. default bin}

\item{plot.it}{logical. Se vuoi che ti disegni il grafico}

\item{soglia}{la soglia desiderata verra tracciata. default 0.5}
}
\value{
a list: i primi 2 elementi servono per la curva lift, gli altri 2 per la curva roc
\item{x1}{vettore, proporzione cumulata dei soggetti previsti}
\item{x2}{fattore di miglioramento, ovvero \code{(n22/n2.)/(n.2/n)}. Ovvero faccio la somma dei soggetti previsti al diminuire della soglia (n22) fratto la proporzione degli 1 (n.2/n) fratto il numero di soggetti che ho classificato come 1 (n2.)}
\item{'1-spec'}{ vettore 1-specificita'}
\item{sens}{ vettore sensitivita'}
}
\description{
disegna la funzione e ti calcola la specificita' e sensitivita'
}
