% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_crossvalidation.R
\name{lm_crossvalidation}
\alias{lm_crossvalidation}
\title{LM cv}
\usage{
lm_crossvalidation(i, formula = NULL, y, x, folds)
}
\arguments{
\item{i}{i-esimo fold}

\item{formula}{formula del modello}

\item{y}{vettore. variabile risposta}

\item{x}{matrice variabili esplicative}

\item{folds}{prodotto da kfold function}
}
\description{
fonisce le previsioni della cv, da utilizzare dentro sapply(1:k,...)
}
