% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_crossvalidation.R
\name{mars_crossvalidation}
\alias{mars_crossvalidation}
\title{GLM cv}
\usage{
mars_crossvalidation(i, formula = NULL, y, x, folds, family = binomial)
}
\arguments{
\item{i}{i-esimo fold}

\item{formula}{formula del modello}

\item{y}{vettore. variabile risposta}

\item{x}{matrice variabili esplicative}

\item{folds}{prodotto da kfold function}

\item{family}{qualsiasi}
}
\description{
fonisce le previsioni della cv, da utilizzare dentro sapply(1:k,...)
}
