% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multifactor.R
\name{multifactor}
\alias{multifactor}
\alias{multifactor.data.frame}
\alias{multifactor.data.table}
\title{Multiple factor}
\usage{
multifactor(data, subset = NULL)

\method{multifactor}{data.frame}(data, subset = NULL)

\method{multifactor}{data.table}(data, subset = NULL)
}
\arguments{
\item{subset}{}
}
\value{
\code{NULL}
}
\description{
convertire più righe contemporaneamente in fattori
}
