% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinumeric.R
\name{multinumeric}
\alias{multinumeric}
\title{multiple numeric}
\usage{
multinumeric(data, subset = NULL)
}
\arguments{
\item{data}{data.frame}

\item{subset}{vector numero colonna o header}
}
\description{
convertire più righe contemporaneamente in numerici
}
