% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_boot.R
\name{my_boot}
\alias{my_boot}
\title{Title}
\usage{
my_boot(data, folds, time1 = 10)
}
\arguments{
\item{data}{data.frame}

\item{folds}{a vector}

\item{time1}{quante volte il campione originale deve essere ingrandito}
}
\description{
Title
}
