% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_boxplot.R
\name{my_boxplot}
\alias{my_boxplot}
\alias{my_boxplot.data.frame}
\alias{my_boxplot.data.table}
\title{Boxplot}
\usage{
my_boxplot(data, subset = "numeric")

\method{my_boxplot}{data.frame}(data, subset = "numeric")

\method{my_boxplot}{data.table}(data, subset = "numeric")
}
\arguments{
\item{subset}{}
}
\description{
Boxplot by ggplot2
}
\examples{

}
