% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mystep.R
\name{mystep}
\alias{mystep}
\title{mystep}
\usage{
mystep(object)
}
\arguments{
\item{object}{modello completo glm}
}
\value{
modello migliore in base all'aic
}
\description{
step un po' più veloce in parallelo, la funzione non crea i cluster in automatico
}
