% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_ratio.R
\name{na_ratio}
\alias{na_ratio}
\title{Proporzione NA}
\usage{
na_ratio(data, table = F, risposta = "y", subset = NULL)
}
\arguments{
\item{subset}{}
}
\description{
Proporzione NA
}
\examples{
# Per affiancare 2 grafici in un unica finestra:
library(gridExtra)
grid.arrange(plot1,p,nrow=1,ncol=2)
data=dati
}
