% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcknn_crossvalidation.R
\name{pcknn_crossvalidation}
\alias{pcknn_crossvalidation}
\title{principal component regression predictions cv}
\usage{
pcknn_crossvalidation(i, y, x, folds, cor = T, npc, K, use.all = F)
}
\arguments{
\item{i}{i-esimo fold}

\item{y}{vettore. variabile risposta}

\item{x}{matrice variabili esplicative}

\item{folds}{prodotto da kfold function}

\item{cor}{logical. if correlation or covariance}

\item{npc}{numero componenti principali}

\item{K}{numero vicini più prossimi}

\item{use.all}{quando più di K punti sono ugualmente vicini, se prenderli tutti o sceglierne K a caso.}
}
\description{
fonisce le previsioni della cv, da utilizzare dentro sapply(1:k,...)
}
