% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.pc.R
\name{plot.pc}
\alias{plot.pc}
\title{plot principal components}
\usage{
\method{plot}{pc}(pc, npc = min(10, length(pc$sdev)), cumsum = F)
}
\arguments{
\item{pc}{object princomp}

\item{npc}{numero pc da plottare}

\item{cumsum}{quale dei 2 grafici}
}
\description{
plot principal components
}
