% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_err.R
\name{plot_err}
\alias{plot_err}
\title{grafico errore di classificazione}
\usage{
plot_err(m, legend = NULL, ylab = "classification error", oob = T)
}
\arguments{
\item{m}{lista oggetti randomforest o ada}

\item{legend}{vettore dei nomi della legenda}
}
\description{
grafico errore di classificazione
}
