% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomforest_crossvalidation.R
\name{randomforest_crossvalidation}
\alias{randomforest_crossvalidation}
\title{random forest cv}
\usage{
randomforest_crossvalidation(i, formula = NULL, y, x, folds, mtry = if
  (!is.null(y) && !is.factor(y)) max(floor(ncol(x)/3), 1) else
  floor(sqrt(ncol(x))))
}
\arguments{
\item{i}{i-esimo fold}

\item{formula}{formula del modello}

\item{y}{vettore. variabile risposta}

\item{x}{matrice variabili esplicative}

\item{folds}{prodotto da kfold function}

\item{mtry}{}
}
\description{
fonisce le previsioni della cv, da utilizzare dentro sapply(1:k,...)
}
