% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relaxlasso cv.R
\name{relaxlasso_crossvalidation}
\alias{relaxlasso_crossvalidation}
\title{relaxlasso_crossvalidation}
\usage{
relaxlasso_crossvalidation(i, y, x, folds, family = "binomial",
  s = "lambda.min")
}
\arguments{
\item{i}{}

\item{y}{response}

\item{x}{matrix or data.frame}

\item{folds}{10}

\item{family}{binomial}

\item{s}{s=c("lambda.1se","lambda.min") default lambda.min}
}
\description{
relaxlasso_crossvalidation
}
