% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riassunto_cat.R
\name{riassunto_cat}
\alias{riassunto_cat}
\alias{riassunto_cat.data.frame}
\alias{riassunto_cat.data.table}
\alias{riassunto_cat.default}
\title{Riassunto delle variabili categoriali}
\usage{
riassunto_cat(data, dropNA = F, freqmod = F, subset = 1:ncol(data),
  order_by = "ncol", ...)

\method{riassunto_cat}{data.frame}(data, dropNA = F, freqmod = F,
  subset = 1:ncol(data), order_by = "ncol", ...)

\method{riassunto_cat}{data.table}(data, dropNA = F, freqmod = F,
  subset = 1:ncol(data), order_by = "ncol", ...)

\method{riassunto_cat}{defaut}(data, dropNA = F, freqmod = F)
}
\arguments{
\item{data}{data.frame}

\item{dropNA}{logical: se vuoi la varianza delle variabili considerando gli NA come categoria.}

\item{freqmod}{logical: se vuoi l'elenco delle categoria e la loro numerosita, e la loro somma cumulata}

\item{subset}{numero delle colonne, nome delle colonne o \code{'factor'}.}
}
\value{
a list: il primo elemento da l'elenco delle categorie, e la somma cumulata da utilizzare quando si decidono di aggregare le categorie.
Il secondo elemento è una matrice contenente la percentuale di valori mancanti, oss*livello e la varianza eliminando le oss mancanti, opzionale la 4 colonna varNA.
}
\description{
Analisi delle variabili categoriali
}
