% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riassunto_corto.R
\name{riassunto_corto}
\alias{riassunto_corto}
\title{Riassunto  corto}
\usage{
riassunto_corto(data, subset, summary = F)
}
\arguments{
\item{data}{}

\item{subset}{}

\item{summary}{}
}
\description{
Riassunto  corto
}
