% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_cat.R
\name{search_cat}
\alias{search_cat}
\title{Ricerca categoria}
\usage{
search_cat(data, cat)
}
\arguments{
\item{data}{data.frame}

\item{cat}{vettore delle categorie}
}
\value{
quali colonne hanno quella categoria
}
\description{
ricerca una o più categorie in un dataset
}
