% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_constant.R
\name{search_constant}
\alias{search_constant}
\title{Trovare quelle variabili con una varianza maggiore della soglia e i fattori con un numero di modalità maggiore di 1 (mettere soglia uguale a 0)}
\usage{
search_constant(data, soglia = 0)
}
\arguments{
\item{data}{data.frame}
}
\description{
Trovare quelle variabili con una varianza maggiore della soglia e i fattori con un numero di modalità maggiore di 1 (mettere soglia uguale a 0)
}
