% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_env.R
\name{search_env}
\alias{search_env}
\title{Search environment
Search an element of environment by name}
\usage{
search_env(name)
}
\arguments{
\item{name}{}
}
\description{
Search environment
Search an element of environment by name
}
