% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_leakers.R
\name{search_leakers}
\alias{search_leakers}
\title{Trovare le variabili leakers}
\usage{
search_leakers(data, soglia = 0.5)
}
\arguments{
\item{data}{data.frame}
}
\value{
una tabella con le variabili che hanno ottenuto r^2 piu' alto sulla risposta e una con le variabili che hanno un tempo eccessivo di lm
}
\description{
cerca le variabili leakers
}
