% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_path.R
\name{search_path}
\alias{search_path}
\title{search_path}
\usage{
search_path(object)
}
\arguments{
\item{object}{glmnet}
}
\value{
i nomi delle variabili nell'ordine in cui sono entrate
}
\description{
search_path
}
