% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_soglia.R
\name{search_soglia}
\alias{search_soglia}
\title{Cercare la soglia}
\usage{
search_soglia(previsti, g, fn)
}
\arguments{
\item{previsti}{mettere le probabilità di 1}

\item{g}{la variabile risposta}

\item{fn}{percentuale di falsi negativi}
}
\description{
Cercare la soglia per avere un determinato numero di falsi negativi
}
