% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_vuote.R
\name{search_vuote}
\alias{search_vuote}
\title{Trovare certe categorie}
\usage{
search_vuote(data, condition)
}
\arguments{
\item{data}{vector or data.frame.}

\item{condition}{grandezza massima della categoria}
}
\value{
il vettore dei nomi delle categorie
}
\description{
Trova le categorie che soddisfano una determinata espressione, utilizzata per vedere quelle con bassa numerosità campionaria. Se è un data.frame guardare le categorie vuote
}
