% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_word.R
\name{search_word}
\alias{search_word}
\title{search word}
\usage{
search_word(word, data, head = T)
}
\arguments{
\item{word}{parola da cercare}

\item{data}{data.frame}

\item{head}{logical: se guardi le prime o le ultime le lettere}
}
\value{
le posizioni corrispondenti
}
\description{
Gli inserisci le iniziali o le finali di una colonna e ti darà le posizioni corrispondenti
}
