% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_varlasso.R
\name{select_varlasso}
\alias{select_varlasso}
\title{selezione variabili}
\usage{
select_varlasso(object, data, s = "lambda.min")
}
\arguments{
\item{object}{cv.glmnet}

\item{data}{data.frame o matrix}

\item{s}{lambda.min o lambda.1se}
}
\description{
selezione variabili di un modello (utilizzato quando si fa il lasso)
}
