% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soglia.R
\name{soglia}
\alias{soglia}
\title{soglia}
\usage{
soglia(a, b, c = 0, d = 0, e = 0, f = 0, g = 0, h = 0, roc = T)
}
\arguments{
\item{a}{oggetto ottenuto dalla funzione tabella.sommario}

\item{roc}{logical. T disegna la soglia della curva ROC,altrimenti quella della curva lift.}
}
\value{
solo i grafici
}
\description{
disegna su un grafico una retta verticale che rappresenta la soglia al 50%
}
