% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_crossvalidation.R
\name{svm_crossvalidation}
\alias{svm_crossvalidation}
\title{LM cv}
\usage{
svm_crossvalidation(i, formula = NULL, y, x, folds, kernel = "linear", cost,
  gamma = 1, degree = 3)
}
\arguments{
\item{i}{i-esimo fold}

\item{formula}{formula del modello}

\item{y}{vettore. variabile risposta}

\item{x}{matrice variabili esplicative}

\item{folds}{prodotto da kfold function}

\item{kernel}{}

\item{cost}{}

\item{gamma}{}

\item{degree}{}
}
\description{
fonisce le previsioni della cv, da utilizzare dentro sapply(1:k,...)
}
