% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabella.sommario.R
\name{tabella.sommario}
\alias{tabella.sommario}
\title{tabella.sommario}
\usage{
tabella.sommario(previsti, osservati, print = T)
}
\arguments{
\item{previsti}{mettere un vettore di 0 o 1 (o T o F)}

\item{osservati}{variabile risposta}
}
\value{
a list
\item{n}{tabella di frequenza tra previsti e osservati}
\item{'1-spec'}{numero. 1-specificita' del modello corrente,ovvero i falsi positivi}
\item{fsp}{frazione soggetti previsti, serve per calcolare la soglia nella curva lift}
}
\description{
funzione che calcola matrice di confusione e gli errori di classificazione
}
