% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabella_riassunto.R
\name{tabella_riassunto}
\alias{tabella_riassunto}
\title{Tabella riassuntiva}
\usage{
tabella_riassunto(e = NULL, previsti, osservati, nome_righe = c("err_tot",
  "falsi_pos", "falsi_neg", "err_media", "auc"))
}
\arguments{
\item{e}{mse}

\item{t}{list delle tabelle da tabella.sommario}

\item{a}{vettore degli auc}

\item{rownames}{nome delle righe}
}
\description{
Tabella riassuntiva
}
