% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_datetime.R
\name{to_datetime}
\alias{to_datetime}
\title{Convert to datetime
convert string or number to POSIXct}
\usage{
to_datetime(data, pattern = NULL, columns = NULL,
  format = "\%Y-\%m-\%d \%H:\%M:\%S")
}
\arguments{
\item{format}{}
}
\description{
Convert to datetime
convert string or number to POSIXct
}
