% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_test_split.R
\name{train_test_split}
\alias{train_test_split}
\title{Title}
\usage{
train_test_split(data, train_size, test_size, seed = 0, stratify = F)
}
\arguments{
\item{data}{}

\item{train_size}{}

\item{test_size}{}

\item{seed}{}

\item{stratify}{}
}
\description{
Title
}
