% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_crossvalidation.R
\name{tree_crossvalidation}
\alias{tree_crossvalidation}
\title{LM cv}
\usage{
tree_crossvalidation(i, formula = "as.factor(y)~.", x, folds, J)
}
\arguments{
\item{i}{i-esimo fold}

\item{formula}{formula del modello}

\item{x}{matrice variabili esplicative}

\item{folds}{prodotto da kfold function}

\item{J}{}
}
\description{
fonisce le previsioni della cv, da utilizzare dentro sapply(1:k,...)
}
