% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vip2.R
\name{vip2}
\alias{vip2}
\title{Variable importance plot}
\usage{
vip2(object, method = 2, max.var.show = 30, cex.names = 0.8,
  cex.axis = 0.8, cex.lab = 0.8, ylim = c(0, mvs), xlab = NULL, ...)
}
\arguments{
\item{object}{random forest object or vector}

\item{method}{1, 2,3. 1= Mean decrease in accuracy, 2=Mean decrease in Gini Index, se l'oggetto non è random forest serve solo come titolo del grafico. 3=Relative importance}
}
\description{
Altra possibilità di grafico rispetto a vip
}
